// src/domain/error/user/user-error-codes.enum.ts
export enum UserErrorCode {
  INVALID_NAME = "INVALID_NAME",
  INVALID_EMAIL = "INVALID_EMAIL",
  INVALID_PHONE = "INVALID_PHONE",
  INCORRECT_PASSWORD = "USER_INCORRECT_PASSWORD",
  CREATION_FAILED = "USER_CREATION_FAILED",
  UPDATE_FAILED = "USER_UPDATE_FAILED",
  PASSWORD_CHANGE_FAILED = "USER_PASSWORD_CHANGE_FAILED",
  GENERIC_USER_ERROR = "USER_GENERIC_ERROR",
  NAME_ERROR = "USER_NAME_ERROR",
  EMAIL_ERROR = "USER_EMAIL_ERROR",
  PHONE_ERROR = "USER_PHONE_ERROR",
  CONFLICT_ERROR = "USER_CONFLICT_ERROR"
}